refactor: extract RLM globals to helper
- Extracted RLM globals (`rlm_shared_*`, `rlm_map`, `rlm_trace`, etc.) from `src/tools/code_exec.ts` to `src/tools/rlm-helpers.ts` to improve code maintainability and testability.
- Updated `code_exec` tool to inject globals using the new helper.
- Verified functional parity and robustness with all tests.
